test(run): run tests and fix imports if needed
