+++ queclink_tramas.py
@@ -1,9 +1,9 @@
-# Shim de compatibilidad para los analizadores de tramas Queclink.
-"""Importa y reexporta funciones públicas desde :mod:`queclink.parser`."""
+"""Shim de compatibilidad para los analizadores de tramas Queclink."""
+
+# DEPRECATION: migrado a queclink.parser
 
 from queclink.parser import parse_gteri, parse_gtinf, parse_line
 
 __all__ = ["parse_line", "parse_gteri", "parse_gtinf"]
 
-# Compatibilidad adicional con el nombre anterior usado en pruebas internas.
 parse_line_to_record = parse_line
